+++ packages/model-adapter/src/index.ts
@@ -1,5 +1,6 @@
 export * from './adapter';
 export * from './client';
+export * from './images';
 export * from './diagnostics';
 export * from './host-cli-transport';
 export * from './run-record';